+++ Sources/StadiaMaps/Models/NearestRoadsRequest.swift
@@ -17,6 +17,7 @@ public struct NearestRoadsRequest: Codable, JSONEncodable, Hashable {
         case instructions
     }
 
+    static let locationsRule = ArrayRule(minItems: 1, maxItems: nil, uniqueItems: false)
     public var units: DistanceUnit?
     public var language: ValhallaLanguages?
     /** The level of directional narrative to include. Locations and times will always be returned, but narrative generation verbosity can be controlled with this parameter. */