+++ Sources/StadiaMaps/Models/RouteRequest.swift
@@ -17,10 +17,23 @@ public struct RouteRequest: Codable, JSONEncodable, Hashable {
         case instructions
     }
 
+    public enum Format: String, Codable, CaseIterable {
+        case json
+        case osrm
+    }
+
+    static let locationsRule = ArrayRule(minItems: 2, maxItems: nil, uniqueItems: false)
     public var units: DistanceUnit?
     public var language: ValhallaLanguages?
     /** The level of directional narrative to include. Locations and times will always be returned, but narrative generation verbosity can be controlled with this parameter. */
     public var directionsType: DirectionsType? = .instructions
+    /** The output response format. The default JSON format is extremely compact and ideal for web or data-constrained use cases where you want to fetch additional attributes on demand in small chunks. The OSRM format is much richer and is configurable with significantly more info for turn-by-turn navigation use cases. */
+    public var format: Format?
+    /** Optionally includes helpful banners with timing information for turn-by-turn navigation. This is only available in the OSRM format. */
+    public var bannerInstructions: Bool?
+    /** Optionally includes voice instructions with timing information for turn-by-turn navigation. This is only available in the OSRM format. */
+    public var voiceInstructions: Bool?
+    public var filters: AnnotationFilters?
     /** An identifier to disambiguate requests (echoed by the server). */
     public var id: String?
     public var locations: [RoutingWaypoint]
@@ -37,10 +50,14 @@ public struct RouteRequest: Codable, JSONEncodable, Hashable {
     /** Determines whether the output should include roundabout exit instructions. */
     public var roundaboutExits: Bool? = true
 
-    public init(units: DistanceUnit? = nil, language: ValhallaLanguages? = nil, directionsType: DirectionsType? = .instructions, id: String? = nil, locations: [RoutingWaypoint], costing: CostingModel, costingOptions: CostingOptions? = nil, excludeLocations: [RoutingWaypoint]? = nil, excludePolygons: [[[Double]]]? = nil, alternates: Int? = nil, elevationInterval: Float? = 0.0, roundaboutExits: Bool? = true) {
+    public init(units: DistanceUnit? = nil, language: ValhallaLanguages? = nil, directionsType: DirectionsType? = .instructions, format: Format? = nil, bannerInstructions: Bool? = nil, voiceInstructions: Bool? = nil, filters: AnnotationFilters? = nil, id: String? = nil, locations: [RoutingWaypoint], costing: CostingModel, costingOptions: CostingOptions? = nil, excludeLocations: [RoutingWaypoint]? = nil, excludePolygons: [[[Double]]]? = nil, alternates: Int? = nil, elevationInterval: Float? = 0.0, roundaboutExits: Bool? = true) {
         self.units = units
         self.language = language
         self.directionsType = directionsType
+        self.format = format
+        self.bannerInstructions = bannerInstructions
+        self.voiceInstructions = voiceInstructions
+        self.filters = filters
         self.id = id
         self.locations = locations
         self.costing = costing
@@ -56,6 +73,10 @@ public struct RouteRequest: Codable, JSONEncodable, Hashable {
         case units
         case language
         case directionsType = "directions_type"
+        case format
+        case bannerInstructions = "banner_instructions"
+        case voiceInstructions = "voice_instructions"
+        case filters
         case id
         case locations
         case costing
@@ -74,6 +95,10 @@ public struct RouteRequest: Codable, JSONEncodable, Hashable {
         try container.encodeIfPresent(units, forKey: .units)
         try container.encodeIfPresent(language, forKey: .language)
         try container.encodeIfPresent(directionsType, forKey: .directionsType)
+        try container.encodeIfPresent(format, forKey: .format)
+        try container.encodeIfPresent(bannerInstructions, forKey: .bannerInstructions)
+        try container.encodeIfPresent(voiceInstructions, forKey: .voiceInstructions)
+        try container.encodeIfPresent(filters, forKey: .filters)
         try container.encodeIfPresent(id, forKey: .id)
         try container.encode(locations, forKey: .locations)
         try container.encode(costing, forKey: .costing)