+++ Sources/StadiaMaps/Models/TraceAttributeFilterOptions.swift
@@ -16,6 +16,7 @@ public struct TraceAttributeFilterOptions: Codable, JSONEncodable, Hashable {
         case exclude
     }
 
+    static let attributesRule = ArrayRule(minItems: 1, maxItems: nil, uniqueItems: false)
     public var attributes: [TraceAttributeKey]
     /** Determines whether the list of attributes will be used as a whitelist or a blacklist. */
     public var action: Action